new.rb: Add capability to create sub-task tickets
Users can now create sub-task tickets for a parent ticket.

How?
The "new" command was expanded to include an optional parent
ticket argument. If not specified, the ticket key is taken
from the git branch. Finally, params is updated to include
a parent key in the fields hash for sub-task tickets.